+++ tests/test_kernel/cutedsl/test_flash_attn_fast.py
@@ -480,7 +480,7 @@ def test_flash_attn_combine(num_splits, seqlen, d, dtype):
     )
     assert_close(
         out,
-        out_ref,
+        out_ref.to(out.dtype),
         atol=2 * (out_pt - out_ref).abs().max().item(),
         rtol=0,
         mismatch_threshold=1e-5,
